Changes to handling of file extraction
File extension list for extraction is configurable in config file. Defaults to: gz, tar, tar.xz, txz, zip
Compressed files are recursively extracted. If a zip contains multiple tar.xz files, the tar.xz files will be extracted as well.
Only run the extract piece, if the file extension is in the list of expected extensions
Change log messages so running the script is more verbose about what is going on. More like the old output, but better.